--- models/User.js
@@ -1,17 +1,0 @@
-const mongoose = require('mongoose')
-
-const UserSchema = new mongoose.Schema({
-    email: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    password: {
-        type: String,
-        required: true,
-
-    }
-})
-
-const User = mongoose.model('User', UserSchema)
-module.exports = User
